backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ML/NLP
spacy==3.7.2
sentence-transformers==2.3.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.0

# GenAI
google-generativeai==0.3.2
groq==0.4.1

# Cache & Queue
redis==5.0.1
celery==5.3.6

# Payments
stripe==7.11.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# For PDF generation (If needed)
reportlab==4.0.9
